Test import data parameters by trying to download 1 record from a table.
Moved the query functions to their own source file for easier code management.
